Move char switch/case to main.c, remove duplicate goto, other misc changes
